new ComplexPlane.cpp + main.cpp

+++ ComplexPlane.h
@@ -27,7 +27,7 @@ private:
     int m_zoomCount;
     float m_aspectRatio;
 
-    int countIterations(sf::Vector2f coord);
+    size_t countIterations(sf::Vector2f coord);
     void iterationsToRGB(size_t count, sf::Uint8& r, sf::Uint8& g, sf::Uint8& b);
     sf::Vector2f mapPixelToCoords(sf::Vector2i mousePixel);
